Move SessionActivationChecker to session package Create isSessionActive method in SessionManager Create getRedErrorMessages method in SessionDocumentReader
