+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/session/SessionManager.java
@@ -14,6 +14,7 @@ public class SessionManager {
 
     private final SessionDocumentReader documentReader;
     private final SessionFacade sessionFacade;
+    private final SessionActivationChecker activationChecker;
 
     public Session getActiveSession() {
         try {
@@ -27,5 +28,9 @@ public class SessionManager {
             throw new SessionException(
                     getClass().getSimpleName() + " can't provide an active Session.", e);
         }
+    }
+
+    public boolean isSessionActive(final WebDocument webDocument) {
+        return activationChecker.isSessionActive(webDocument);
     }
 }
